Created Forms Folder, fixed EditForm props
